@@ -1,20 +1,17 @@
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 31
+    buildToolsVersion '30.0.3'
 
     defaultConfig {
         applicationId "com.itmd2"
         minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 9
+        targetSdkVersion 31
+        versionCode 10
         versionName "1.0"
-
-        manifestPlaceholders += ["fabric-api-key": readProperty("FABRIC_API_KEY")]
     }
 
     buildTypes {
