@@ -1,30 +1,17 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.50'
+    ext.kotlin_version = '1.5.0'
     repositories {
-        jcenter()
         google()
-        maven { url 'https://maven.fabric.io/public' }
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0-beta04'
-        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
+        classpath 'com.android.tools.build:gradle:7.0.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'io.fabric.tools:gradle:1.25.4'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
-    }
-}
-
-allprojects {
-    repositories {
-        jcenter()
-        maven { url "https://jitpack.io" }
-        maven { url "https://maven.google.com" }
-        maven { url 'https://maven.fabric.io/public' }
-        google()
     }
 }
 
@@ -34,14 +21,17 @@ task clean(type: Delete) {
 
 ext {
     espressoVersion = "2.2.2"
-    archLifecycleVersion = "1.1.1"
-    archRoomVersion = "1.1.1"
-    support_version = '27.1.1'
+    archLifecycleVersion = "2.4.0"
+    archRoomVersion = "2.3.0"
 }
 
-Object readProperty(String propertyName) {
-    def props = new Properties()
-    file("keys.properties").withInputStream { props.load(it) }
-
-    return props.getProperty(propertyName)
+allprojects {
+    repositories {
+        repositories {
+            google()
+            mavenCentral()
+            // for photo view
+            maven { url "https://jitpack.io" }
+        }
+    }
 }
