@@ -1,9 +1,9 @@
 package com.itmoldova.adapter
 
 import android.content.res.Resources
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
 
 import com.itmoldova.list.ArticlesFragment
 import com.itmoldova.model.Category
