@@ -1,7 +1,7 @@
 package com.itmoldova.comments
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 import com.itmoldova.R
 
