@@ -1,7 +1,7 @@
 package com.itmoldova.util
 
 import android.content.Context
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import com.itmoldova.R
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
