@@ -1,12 +1,12 @@
 package com.itmoldova.adapter
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.itmoldova.R
 import com.itmoldova.model.Article
 import com.itmoldova.util.UiUtils
