@@ -1,7 +1,7 @@
 package com.itmoldova.db
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.RoomDatabase
 import com.itmoldova.model.Article
 
 @Database(entities = arrayOf(Article::class), version = 1)
